Fix garbage/undefined load on E_Message_Server_IP
SOCK_ADDR_SIZE can't be used here because server_sockaddr. The
structure is uninitialized so we can't look up the ss_family.

It's just fine calling getsockname with sizeof the full
sockaddr_storage structure, the macro is overkill. I wonder if that's
the case in other places.

Caught by clang-analyzer.