Generate manual as part of the build process
sipp.1 is auto-generated and is prone to getting out of date or
having features it does not support per ./configure options. Make it
part of the build process.

This change also results in the manual page for sipp to get installed
in a normal place ($prefix/man/man1/). A more useful short description
is also included (previously it was "manual page for sipp").