Remove auto-generated manual page
sipp.1 is now built during make (provided that help2man is installed).
This ensures that the manual page reflects the features implemented
in the sipp program. For this reason, do not distribute a pre-built
manual page either.

auto-generate-files.sh is not really useful anymore and removed. It
did not build sipp anyway before generating a man page, possibly
leading to an out-of-date manual page.