Fix invalid authentication response calculation
The digest calculation when "qop" is given follows from:

    HA1=MD5(username:realm:password)
    HA2=MD5(method:digestURI)
    response=MD5(HA1:nonce:nonceCount:clientNonce:qop:HA2)

When qop is not given, the response is:

    response=MD5(HA1:nonce:HA2)

Currently the qop parameter is hard-coded to "auth-int" which fails to
compute the correct hash for qop=auth. Fix this by using the correct
authentication type instead.

References:

 - https://tools.ietf.org/html/rfc3261#section-22.4
 - https://en.wikipedia.org/wiki/Digest_access_authentication#Overview

Fixes https://github.com/SIPp/sipp/issues/87