Rewrite get_remote_media_addr to be AF-independent
Ported to getaddrinfo (with AI_NUMERICHOST and AI_NUMERICSERV to avoid
unnecessary DNS lookups). Simplified parsing and removes a lot of
temporary allocations as well.

Adds a simple mock call class for testing with some special helper
methods to test class internals. This is not really an ideal was to
test, but good enough for now until more refactoring can be done.