Fix bad pcap parsing
Looks like the ipv4 packet length code is completely wrong.

The original code seems to be doing special logic on OSX, Cygwin and
FreeBSD, then again on HP UX and then again on Linux (and everything
else) because of, what I can only guess, must be differences in
structure padding. Mark the iphdr struct as packed for good measure.

There is no reason for the + 4, it shouldn't have slipped in. This
should fix #223.

Also fixes parsing ipv6 pcaps. The reliance on ethhdr is wrong.