SSL: separate CA file
This commit allows using a separate file to specify the CA to trust.
This way we have two files (cert, key) for our own certificate and
one file with the trusted CA.  Putting the own certificate and the CA
in the same file, as previously done with the tls_cert option, does
not seem to work.