Increment nonce count and get auth method from message being sent.
Nonce count changes are based on the patch described here:

https://sipp-users.narkive.com/6Gl681EP/patch-to-correctly-increment-nc-in-authorization-header

The old way of getting the auth method from the CSeq of the last header would fail if the previous message sent was a different method than the current one. It seems simple and safe to just use the method of the message being sent.