Rewrite DTD based on docs
I attempted writing my own custom scenarios using the DTD as a guide,
but found that there were several errors in it. This patch re-writes
the DID based on the docs found on [Read The Docs][1].

There's still some improvements that can happen, but this seems to work
much better.

 [1]: https://sipp.readthedocs.io/en/v3.6.0/scenarios/ownscenarios.html