Socket: Reorder abort operations on read/write error
In 7c7c713971193a83399be71fe129b0f9c5f8864c the order of abort() and
sockets_pending_reset.insert(socket) was changed. Starting from this
commit, abort() is called before inserting to the list.

If the refcount drops to 0, the socket erases itself from the list (in
which it still was not inserted), and then deletes itself. After that,
the deleted socket is inserted to the list.

Fixed by reverting to the original order.