Consistently unescape XML attributes when loading scenario
This changes the behaviour of `xp_get_string()` in order to
systematically unescape the returned values, as other XML parsers do.
The custom unescaping in the handling of the `<ereg>` element has been
removed as it is now unneeded.

This will make it possible to e.g. compare strings that contain special
XML characters using `<strcmp>`.

A regression test case has been added to highlight the problem.

https://github.com/SIPp/sipp/issues/458