Fix version.h include in release tar (version.h should be in include/)
version.h will now be searched first in the build dir, then in the
include. If it's only found in the latter, that one is used.

(When buidling to a directory other than project root, like debhelper
does, the version.h was previously not found because the it searches the
build dir, but not the project root.)

This also reinstates the stub version.h that #errors when you have
downloaded a git snapshot.