Some snprintf/strcat fixage in auth module
Yes. I know fully well that this is not safe snprintf usage as it can
return negative values or values that are larger than result_len. Feel
free to file a PR.

Previously, this was done with even less checks and large enough
buffers. The buffers are still large enough, and the written+= counts
saves us the strcat.